prow/agent-workflows/agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ack-codegen-agent"
version = "0.1.0"
description = "AWS Controllers for Kubernetes (ACK) Code Generation Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Amazon Web Services", email = "aws-controllers-k8s@amazon.com"},
]
dependencies = [
    "strands-agents>=0.1.3",
    "boto3>=1.28.0",
    "gitpython>=3.1.30",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=14.0.0",
    "psutil>=7.0.0",
    "strands-agents-tools[mem0-memory]>=0.1.3",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["ack_builder_agent"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
